requirements.txt
discord.py
praw

